Move misc linop definitions into new module
